backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai>=1.82.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
psutil==5.9.5
yfinance==0.2.18
